Merge with dev add approve invoice

See merge request propensi-2023-2024-genap/kelas-a/a04-megapro!45

+++ invoicesync/src/main/java/com/megapro/invoicesync/model/ApprovalFlow.java
@@ -21,7 +21,7 @@ public class ApprovalFlow {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long approvalRank;
+    private int approvalRank;
     private String approverRole;
     private Long nominalRange;
 }
